Remove references to user primary email
